backend/requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aioice==0.10.2
aiortc==1.14.0
aiosignal==1.4.0
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asyncpg==0.31.0
attrs==26.1.0
av==16.1.0
bcrypt==5.0.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
cryptography==46.0.6
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
email-validator==2.3.0
eval_type_backport==0.3.1
fastapi==0.135.2
filelock==3.25.2
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.2.0
google-auth==2.49.1
google-crc32c==1.8.0
google-genai==1.68.0
googleapis-common-protos==1.73.1
greenlet==3.3.2
groq==1.1.2
grpcio==1.80.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.8.0
humanfriendly==10.0
idna==3.11
ifaddr==0.2.0
importlib_metadata==8.7.1
jiter==0.13.0
joblib==1.5.3
livekit==1.1.3
livekit-agents==1.5.1
livekit-api==1.1.0
livekit-blingfire==1.1.0
livekit-plugins-cartesia==1.5.1
livekit-plugins-openai==1.5.1
livekit-plugins-silero==1.5.1
livekit-protocol==1.1.3
llvmlite==0.46.0
loguru==0.7.3
Mako==1.3.10
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
nest-asyncio==1.6.0
nltk==3.9.4
numba==0.64.0
numpy>=2.0.0,<2.3
onnxruntime==1.23.2
openai==2.30.0
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
packaging==26.0
passlib==1.7.4
pillow==12.1.1
pipecat-ai==0.0.107
prometheus_client==0.24.1
propcache==0.4.1
protobuf==5.29.6
psutil==7.2.2
psycopg2-binary==2.9.11
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.1
Pygments==2.19.2
PyJWT==2.12.1
pylibsrtp==1.0.0
pyloudnorm==0.2.0
pyOpenSSL==26.0.0
python-dotenv==1.2.2
python-multipart==0.0.22
PyYAML==6.0.3
regex==2026.2.28
requests==2.33.0
resampy==0.4.3
rich==14.3.3
safetensors==0.7.0
scipy>=1.11.0,<1.16
shellingham==1.5.4
sniffio==1.3.1
soxr==1.0.0
SQLAlchemy==2.0.48
starlette==1.0.0
sympy==1.14.0
tenacity==9.1.4
tokenizers==0.22.2
tqdm==4.67.3
transformers==5.3.0
typer==0.24.1
types-protobuf==6.32.1.20260221
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
watchfiles==1.1.1
websockets<16.0
yarl==1.23.0
zipp==3.23.0
sib-api-v3-sdk==3.0.0
